Move Identity models into Account subfolder.
+ Rename some of the scaffolded Identity classes.

Replace MSTest with xUnit.
+ Add FakeHttpContext package to mock HttpContext easily.
+ Add AutoFixture to help us generate values for our unit tests.
