+++ Project.Tests/Controllers/HomeControllerTest.cs
@@ -3,15 +3,15 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Web.Mvc;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Project;
 using Project.Controllers;
+using Xunit;
 
 namespace Project.Tests.Controllers {
-    [TestClass]
+
     public class HomeControllerTest {
 
-        [TestMethod]
+        [Fact]
         public void Index() {
             // Arrange
             HomeController controller = new HomeController();
@@ -20,10 +20,10 @@ namespace Project.Tests.Controllers {
             ViewResult result = controller.Index() as ViewResult;
 
             // Assert
-            Assert.IsNotNull(result);
+            Assert.NotNull(result);
         }
 
-        [TestMethod]
+        [Fact]
         public void About() {
             // Arrange
             HomeController controller = new HomeController();
@@ -32,10 +32,10 @@ namespace Project.Tests.Controllers {
             ViewResult result = controller.About() as ViewResult;
 
             // Assert
-            Assert.AreEqual("Your application description page.", result.ViewBag.Message);
+            Assert.Equal("Your application description page.", result.ViewBag.Message);
         }
 
-        [TestMethod]
+        [Fact]
         public void Contact() {
             // Arrange
             HomeController controller = new HomeController();
@@ -44,7 +44,7 @@ namespace Project.Tests.Controllers {
             ViewResult result = controller.Contact() as ViewResult;
 
             // Assert
-            Assert.IsNotNull(result);
+            Assert.NotNull(result);
         }
     }
 }
