+++ Project.Tests/packages.config
@@ -1,10 +1,19 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="AutoFixture" version="4.2.0" targetFramework="net47" />
+  <package id="AutoFixture.Xunit2" version="4.2.0" targetFramework="net47" />
+  <package id="FakeHttpContext" version="0.4.1" targetFramework="net47" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.4" targetFramework="net47" />
   <package id="Microsoft.AspNet.Razor" version="3.2.4" targetFramework="net47" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.4" targetFramework="net47" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net47" />
-  <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net47" />
-  <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net47" />
   <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net47" />
+  <package id="xunit" version="2.3.1" targetFramework="net47" />
+  <package id="xunit.abstractions" version="2.0.1" targetFramework="net47" />
+  <package id="xunit.analyzers" version="0.7.0" targetFramework="net47" />
+  <package id="xunit.assert" version="2.3.1" targetFramework="net47" />
+  <package id="xunit.core" version="2.3.1" targetFramework="net47" />
+  <package id="xunit.extensibility.core" version="2.3.1" targetFramework="net47" />
+  <package id="xunit.extensibility.execution" version="2.3.1" targetFramework="net47" />
+  <package id="xunit.runner.visualstudio" version="2.3.1" targetFramework="net47" developmentDependency="true" />
 </packages>
