+++ Project/Models/Account/AccountViewModels.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace Project.Models
+namespace Project.Models.Account
 {
     public class ExternalLoginConfirmationViewModel
     {
